refactored commands, reinstated drive default command
