+++ src/main/kotlin/frc/team6502/robot/commandgroups/VisionAlign.kt
@@ -1,7 +1,7 @@
 package frc.team6502.robot.commandgroups
 
 import edu.wpi.first.wpilibj.command.CommandGroup
-import frc.team6502.robot.commands.auto.RamseteFollowPath
+import frc.team6502.robot.commands.drive.RamseteFollowPath
 import frc.team6502.robot.commands.vision.CollectVisionData
 import frc.team6502.robot.commands.vision.GenerateVisionSpline
 import frc.team6502.robot.commands.vision.SetLEDRing
