+++ src/main/kotlin/frc/team6502/robot/commands/drive/DefaultDrive.kt
@@ -1,4 +1,4 @@
-package frc.team6502.robot.commands.defaults
+package frc.team6502.robot.commands.drive
 
 import edu.wpi.first.wpilibj.Timer
 import edu.wpi.first.wpilibj.command.PIDCommand
