+++ src/main/kotlin/frc/team6502/robot/commands/manip/IntakeCargo.kt
@@ -1,4 +1,4 @@
-package frc.team6502.robot.commands
+package frc.team6502.robot.commands.manip
 
 import edu.wpi.first.wpilibj.Timer
 import edu.wpi.first.wpilibj.command.Command
