+++ src/main/kotlin/frc/team6502/robot/commands/manip/SetElevatorHeight.kt
@@ -1,4 +1,4 @@
-package frc.team6502.robot.commands
+package frc.team6502.robot.commands.manip
 
 import edu.wpi.first.wpilibj.command.InstantCommand
 import frc.team6502.kyberlib.util.units.Length
